# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Geolocal",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/problemes": {
            "get": {
                "tags": [
                    "Problemes"
                ],
                "summary": "liste des problemes",
                "description": "Elle renvoit aussi le nombre total de problemes",
                "operationId": "",
                "requestBody": {
                    "description": "Vous n'avez pas besoin de vous connecter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/problemes/commune/{id}": {
            "get": {
                "tags": [
                    "Problemes"
                ],
                "summary": "Recherche les problemes d'une commune",
                "description": "Renvoie les problemes d'une commune",
                "operationId": "",
                "requestBody": {
                    "description": "Vous n'avez pas besoin de vous connecter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/problemes/category/{id}": {
            "get": {
                "tags": [
                    "Problemes"
                ],
                "summary": "recherche les problemes appartenant a une category",
                "description": "Renvoie les problemes d'une category",
                "operationId": "",
                "requestBody": {
                    "description": "Vous n'avez pas besoin de vous connecter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/problemes/users/{users}": {
            "get": {
                "tags": [
                    "Problemes"
                ],
                "summary": "recherche les problemes d'un utilisateur",
                "description": "Renvoie les problemes propre a un utilisateur",
                "operationId": "",
                "requestBody": {
                    "description": "Vous n'avez pas besoin de vous connecter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/problemes/": {
            "post": {
                "tags": [
                    "Problemes"
                ],
                "summary": "liste des problemes",
                "description": "",
                "operationId": "",
                "requestBody": {
                    "description": "Vous n'avez pas besoin de vous connecter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "longitude",
                                    "latitude",
                                    "image",
                                    "commune",
                                    "arrondissement"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "automatique"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "un text"
                                    },
                                    "audio": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "users": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "NameUser"
                                    },
                                    "anonymes": {
                                        "type": "boolean",
                                        "format": "True|False",
                                        "example": "True"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "longitude": {
                                        "type": "float",
                                        "format": "float",
                                        "example": "15.6"
                                    },
                                    "latitude": {
                                        "type": "float",
                                        "format": "float",
                                        "example": "15.6"
                                    },
                                    "commune_id": {
                                        "type": "int",
                                        "format": "",
                                        "example": "1"
                                    },
                                    "arrondissement_id": {
                                        "type": "int",
                                        "format": "",
                                        "example": "1"
                                    },
                                    "category_id": {
                                        "type": "int",
                                        "format": "",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/problemes/{id}": {
            "delete": {
                "tags": [
                    "Problemes"
                ],
                "summary": "Suppression",
                "description": "Vous devez vous connectez en admin pour effectuer la suppression ",
                "operationId": "",
                "requestBody": {
                    "description": "Vous devez vous connectez d'abord et etre admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Problemes"
                ],
                "summary": "modificatins",
                "description": "Pour modifier un attributs des problemes vous devez etre admin ou admin_locaux",
                "operationId": "",
                "requestBody": {
                    "description": "Vous devez vous connectez d'abord et etre admin ou admin_locaux",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}